backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-chroma
langchain-community
langgraph
chromadb
networkx
pydantic
python-dotenv
